+++ mod/src/main/java/basemod/patches/com/megacrit/cardcrawl/characters/AbstractPlayer/MaxHandSizePatch.java
@@ -12,6 +12,7 @@ import com.megacrit.cardcrawl.actions.unique.DiscoveryAction;
 import com.megacrit.cardcrawl.actions.unique.ExhumeAction;
 import com.megacrit.cardcrawl.actions.unique.SkillFromDeckToHandAction;
 import com.megacrit.cardcrawl.actions.utility.DrawPileToHandAction;
+import com.megacrit.cardcrawl.actions.watcher.ForeignInfluenceAction;
 import com.megacrit.cardcrawl.cards.AbstractCard;
 import com.megacrit.cardcrawl.cards.CardGroup;
 import com.megacrit.cardcrawl.cards.purple.Scrawl;
@@ -77,6 +78,10 @@ import java.util.HashMap;
 @SpirePatch(
 		clz=Scrawl.class,
 		method="use"
+)
+@SpirePatch(
+		clz=ForeignInfluenceAction.class,
+		method="update"
 )
 public class MaxHandSizePatch
 {